Renamed DB stuff, fixed injection vulnerability
