--- chess-website/package.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.9.3",
-    "react-redux": "^8.0.5",
-    "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.4.2"
-  }
-}
